Send slider values only on change
Prior to this patch, slider values would be sent on every loop. This is
unnecessary and can make initial setup difficult (e.g. when using the
serial monitor to verify outputs and correct pot behaviour).

This patch introduces a `lastSliderValues` array to track the last
slider values that were sent to the PC and a `MIN_CHANGE` constant to set the
minimum change required before sending the new slider values.